% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_edge.R
\name{addEdge}
\alias{addEdge}
\title{Add an edge}
\usage{
addEdge(g, nodeStart, nodeDest, attributes)
}
\arguments{
\item{g}{The input graph}

\item{nodeStart}{The first endpoint of the edge}

\item{nodeDest}{The second endpoint of the edge}

\item{attributes}{The attributes to assign to the edge}
}
\value{
The graph, with the added edge
}
\description{
Add an edge
}
