% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_node.R
\name{addNode}
\alias{addNode}
\title{Add a node with assigned layer and attributes to a graph}
\usage{
addNode(g, nodeName, layerName, attributes = NA)
}
\arguments{
\item{g}{The input graph.}

\item{nodeName}{The name of the node to add.}

\item{layerName}{The name of the layer to be assigned to the node.}

\item{attributes}{The attributes of the node to add. This argument must be a named list.}
}
\value{
The mully graph, with the new node.
}
\description{
Add a node with assigned layer and attributes to a graph
}
\examples{
g=mully::demo()
attributes=list("specie"="Homo Sapiens")
addNode(g = g,nodeName = "g3",layerName = "Gene",attributes = attributes)
}
