% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_demo.R
\name{demo}
\alias{demo}
\title{A demo function to test the package}
\usage{
demo(name = "MyFirstMully")
}
\arguments{
\item{name}{Optional name for the graph to be created. By default set to "MyFirstMully"}
}
\value{
A mully test graph with few layers, nodes and edges
}
\description{
A demo function to test the package
}
