% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_export.R
\name{exportCSV}
\alias{exportCSV}
\title{Export mully into CSV files}
\usage{
exportCSV(g, target)
}
\arguments{
\item{g}{The input graph}

\item{target}{The target file in which the files will be generated.}
}
\value{
No return value. Exports the given graph into different CSV files saved in the target folder
}
\description{
Export mully into CSV files
}
\examples{
\dontrun{
g=mully::demo()
exportCSV(g,directory)
}
}
