% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_edge.R
\name{getIDEdge}
\alias{getIDEdge}
\title{Get the ids of the edges connecting two nodes}
\usage{
getIDEdge(g, nodeStart, nodeDest)
}
\arguments{
\item{g}{The input graph}

\item{nodeStart}{The first endpoint of the edge}

\item{nodeDest}{The second endpoint of the edge}
}
\value{
A list containing the ids of the edges connecting the nodes
}
\description{
Get the ids of the edges connecting two nodes
}

