% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_node.R
\name{getIDNode}
\alias{getIDNode}
\title{Get the id of a node}
\usage{
getIDNode(g, nameNode)
}
\arguments{
\item{g}{The input graph}

\item{nameNode}{The name of the node}
}
\value{
The id of the specified node
}
\description{
Get the id of a node
}

