% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_layer.R
\name{getLayer}
\alias{getLayer}
\title{Get the nodes on a layer in a graph}
\usage{
getLayer(g, nameLayer)
}
\arguments{
\item{g}{The input graph.}

\item{nameLayer}{The name of the layer.}
}
\value{
A List of the nodes on the given layer.
}
\description{
Get the nodes on a layer in a graph
}

