% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_layer.R
\name{getLayersCount}
\alias{getLayersCount}
\title{Get the number of layers in a graph}
\usage{
getLayersCount(g)
}
\arguments{
\item{g}{The input graph.}
}
\value{
The count of the layers.
}
\description{
Get the number of layers in a graph
}

