% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_node.R
\name{getNodeAttributes}
\alias{getNodeAttributes}
\title{Get the attributes of a node}
\usage{
getNodeAttributes(g, nameNode)
}
\arguments{
\item{g}{The input graph}

\item{nameNode}{The name of the node}
}
\value{
A framework containing the attributes of the specified node
}
\description{
Get the attributes of a node
}

