% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_import.R
\name{importEdgesCSV}
\alias{importEdgesCSV}
\title{Import Edges to a mully graph from a CSV file}
\usage{
importEdgesCSV(g, file)
}
\arguments{
\item{g}{The mully graph to which the nodes will be added. The graph should already have the layers and the nodes.}

\item{file}{The path to the CSV file containing the edges' information}
}
\value{
The mully graph with the added edges
}
\description{
Import Edges to a mully graph from a CSV file
}
