% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_constructor.R
\name{is.mully}
\alias{is.mully}
\title{Is this a mully graph?}
\usage{
is.mully(g)
}
\arguments{
\item{g}{The input graph}
}
\value{
A boolean whether the graph is or not a mully object
}
\description{
Is this a mully graph?
}
