% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_layer.R
\name{isLayer}
\alias{isLayer}
\title{Verify if the layer exists in a graph}
\usage{
isLayer(g, name)
}
\arguments{
\item{g}{The input graph.}

\item{name}{The name of the layer.}
}
\value{
A boolean value.
}
\description{
Verify if the layer exists in a graph
}
\examples{
g = mully("MyFirstMully",direct = FALSE)
g = addLayer(g, c("Gene", "Drug", "Disease"))
isLayer(g,"Drug")
}
