% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_merge.R
\name{merge}
\alias{merge}
\title{Merge or unite two graphs}
\usage{
merge(g1, g2)
}
\arguments{
\item{g1}{The first graph to merge. This is the base of the merge.}

\item{g2}{The second graph to merge. All of its elements are added to the first graph.}
}
\value{
The merge of the two graphs. The merge is based on the first given graph
}
\description{
Merge or unite two graphs
}
