% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully.R
\docType{package}
\name{mully-package}
\alias{mully}
\alias{mully-package}
\title{The mully package}
\description{
R package to create, modify and visualize graphs with multiple layers.
}
\section{Introduction}{

}

\section{Creating graphs}{

}

\section{Visualization}{

}

\section{Further information}{

}

