% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully.R
\name{mully}
\alias{mully}
\title{Create an empty multilayered graph}
\usage{
mully(name = NA, direct = TRUE)
}
\arguments{
\item{name}{The name to be assigned to the graph.}

\item{direct}{A boolean value, if the graph is directed or not. By default TRUE.}
}
\value{
The created multilayered graph.
}
\description{
Create an empty multilayered graph
}

