% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_visualization.R
\name{plot.mully.3d}
\alias{plot.mully.3d}
\title{Plot the graph in 3D using rgl}
\usage{
\method{plot}{mully.3d}(g)
}
\arguments{
\item{g}{The input graph}
}
\description{
Plot the graph in 3D using rgl
}
