% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_visualization.R
\name{plot.mully}
\alias{plot.mully}
\title{Plot the graph in 2D}
\usage{
\method{plot}{mully}(x, layout, ...)
}
\arguments{
\item{x}{The input graph}

\item{layout}{The layout. Can either be random or scaled}

\item{...}{Other arguments to be passed to \link[igraph]{plot.igraph}}
}
\value{
No return value. A 2D plot of the given graph is created.
}
\description{
Plot the graph in 2D
}
\examples{
g=mully::demo()
plot(g,"Scaled")
}
