% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_visualization.R
\name{plot3d.mully}
\alias{plot3d.mully}
\title{Plot the graph in 3D using rgl}
\usage{
\method{plot3d}{mully}(
  g,
  layers = T,
  vertex.shape = "circle",
  vertex.label = NA,
  edge.width = 1
)
}
\arguments{
\item{g}{The input graph}

\item{layers}{A boolean whether to add the layers or not}
}
\description{
Plot the graph in 3D using rgl
}
\note{

}
