% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_constructor.R
\name{print.mully}
\alias{print.mully}
\title{Print function}
\usage{
\method{print}{mully}(x, ...)
}
\arguments{
\item{x}{The input graph}

\item{...}{Other arguments to be passed to \code{print}}
}
\value{
No return value. The given graph is printed
}
\description{
Print function
}
\examples{
g=mully::demo()
print(g)
}
