% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_edge.R
\name{removeEdge}
\alias{removeEdge}
\title{Delete an edge}
\usage{
removeEdge(g, nodeStart, nodeDest, attributes = NA, multi = FALSE)
}
\arguments{
\item{g}{The input graph}

\item{nodeStart}{The first endpoint of the edge}

\item{nodeDest}{The second endpoint of the edge}

\item{attributes}{The attributes of the edge to delete. Required if the nodes are multi-connected}

\item{multi}{A boolean. Specifies whether to delete multiple edges or not, in case they exist.}
}
\value{
The mully graph with the deleted edges
}
\description{
Delete an edge
}
\examples{
g=mully::demo()
removeEdge(g,"dr1","d2",multi=TRUE)
}
