% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_layer.R
\name{removeLayer}
\alias{removeLayer}
\title{Delete a layer or a set of layers from a graph}
\usage{
removeLayer(g, name, trans = FALSE)
}
\arguments{
\item{g}{The input graph.}

\item{name}{The name or the list of the names of the layers to be deleted.}

\item{trans}{A boolean whether to insert transitive edges or not}
}
\value{
The mully graph, with the given layer and its corresponding nodes and edges removed.
}
\description{
Delete a layer or a set of layers from a graph
}
\examples{
g = mully::demo()
removeLayer(g,"gene",trans=TRUE)
}
