% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_node.R
\name{getNodeAttributes}
\alias{getNodeAttributes}
\title{Get the attributes of a node}
\usage{
getNodeAttributes(g, nameNode, layerByName = FALSE)
}
\arguments{
\item{g}{The input graph}

\item{nameNode}{The name of the node}

\item{layerByName}{A boolean to specify whether to export the layers by name or by ID}
}
\value{
A dataframe containing the attributes of the specified node
}
\description{
Get the attributes of a node
}
\examples{
g=mully::demo()
getNodeAttributes(g,layerByName = TRUE)
}
