% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_visualization.R
\name{plot.mully}
\alias{plot.mully}
\title{Plot the graph in 2D}
\usage{
\method{plot}{mully}(g, layout)
}
\arguments{
\item{g}{The input graph}

\item{layout}{The layout. Can either be random or scaled}
}
\description{
Plot the graph in 2D
}
