% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mully_visualization.R
\name{plot3d.mully}
\alias{plot3d.mully}
\title{Plot the graph in 3D using rgl}
\usage{
\method{plot3d}{mully}(g)
}
\arguments{
\item{g}{The input graph}
}
\description{
Plot the graph in 3D using rgl
}
